# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ypk btpit",
        "description": "Ypk btpit",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/byYpk/{ypk}": {
            "get": {
                "tags": [
                    "Товары"
                ],
                "summary": "Получить товары по YPK",
                "operationId": "42e28fdab38fde60e58f7405177b27bf",
                "parameters": [
                    {
                        "name": "ypk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "ypkId": {
                                                "type": "string"
                                            },
                                            "productName": {
                                                "type": "string"
                                            },
                                            "productCost": {
                                                "type": "number"
                                            },
                                            "productInfo": {
                                                "type": "string"
                                            },
                                            "isProduct": {
                                                "type": "boolean"
                                            },
                                            "photoPath": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "photoUrl": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "statusProductId": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/{product}": {
            "get": {
                "tags": [
                    "Товары"
                ],
                "summary": "Получить товар по ID",
                "operationId": "687fb099aa12137c8ba13c343730bdd9",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "ypkId": {
                                            "type": "string"
                                        },
                                        "productName": {
                                            "type": "string"
                                        },
                                        "productCost": {
                                            "type": "number"
                                        },
                                        "productInfo": {
                                            "type": "string"
                                        },
                                        "isProduct": {
                                            "type": "boolean"
                                        },
                                        "photoPath": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "photoUrl": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "statusProductId": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Товары"
                ],
                "summary": "Удалить товар по ID",
                "operationId": "8076e8fe60c38065603af989523ab5df",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "Товары"
                ],
                "summary": "Создать товар",
                "operationId": "f6c6fa48346dd08aae3060e1e8765913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ProductName": {
                                        "type": "string"
                                    },
                                    "YpkId": {
                                        "type": "string"
                                    },
                                    "StatusProductId": {
                                        "type": "string"
                                    },
                                    "ProductCost": {
                                        "type": "string"
                                    },
                                    "ProductInfo": {
                                        "type": "string"
                                    },
                                    "IsProduct": {
                                        "type": "boolean"
                                    },
                                    "Photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/update": {
            "put": {
                "tags": [
                    "Товары"
                ],
                "summary": "Обновить товар по ID",
                "operationId": "88e635d1eee982a79655ee22cd22d842",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "ProductName": {
                                        "type": "string"
                                    },
                                    "YpkId": {
                                        "type": "string"
                                    },
                                    "StatusProductId": {
                                        "type": "string"
                                    },
                                    "ProductCost": {
                                        "type": "string"
                                    },
                                    "ProductInfo": {
                                        "type": "string"
                                    },
                                    "IsProduct": {
                                        "type": "boolean"
                                    },
                                    "Photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Обновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}